last commit, giving up. intervaltimer broken

+++ src/crc.cpp
@@ -1,13 +1,13 @@
 #include "main.h"
 
-uint8_t CRCTable[256];
+uint_fast8_t CRCTable[256];
 
 void CRCInit(){
-    uint8_t remainder;
+    uint_fast8_t remainder;
     for (int dividend = 0; dividend < 256; ++dividend)
     {
         remainder = dividend << (WIDTH - 8);
-        for (uint8_t bit = 8; bit > 0; --bit)
+        for (uint_fast8_t bit = 8; bit > 0; --bit)
         {
             if (remainder & TOPBIT)
             {
@@ -22,17 +22,17 @@ void CRCInit(){
     }
 }
 
-uint8_t CRCFast(uint8_t msg[], uint8_t nBytes){
-    uint8_t data;
-    uint8_t remainder = 0;
-    for (int byte = 0; byte < nBytes; ++byte)
+uint8_t CRCFast(uint8_t msg[], uint_fast8_t nBytes){
+    uint_fast8_t data;
+    uint_fast8_t remainder = 0;
+    for (uint_fast8_t byte = 0; byte < nBytes; ++byte)
     {
         data = msg[byte] ^ (remainder >> (WIDTH - 8));
         remainder = CRCTable[data] ^ (remainder << 8);
     }
-    return (remainder);
+    return uint8_t(remainder);
 }
 
-ErrorCode CRCCheck(uint8_t msg[], uint8_t nBytes){
+ErrorCode CRCCheck(uint_fast8_t msg[], uint_fast8_t nBytes){
     return NO_ERROR;
 }
